backend/requirements.txt
asgiref==3.2.9
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
cloudinary==1.21.1
dj-database-url==0.5.0
Django==3.0.7
django-cloudinary-storage==0.2.3
django-cors-headers==3.4.0
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
more-itertools==8.4.0
Pillow==7.1.2
pluggy==0.6.0
psycopg2==2.8.5
py==1.8.2
PyJWT==1.7.1
pytest==3.6.0
pytest-django==3.9.0
python-decouple==3.3
python-magic==0.4.18
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
